Split shared plugin
