+++ plugins/database/src/main/kotlin/fr/pickaria/database/models/BankAccounts.kt
@@ -1,6 +1,5 @@
-package fr.pickaria.shared.models
+package fr.pickaria.database.models
 
-import fr.pickaria.shared.BukkitLogger
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.transactions.transaction
