+++ plugins/economy/build.gradle.kts
@@ -3,6 +3,8 @@ repositories {
 }
 
 dependencies {
+	compileOnly(project(":lib"))
+	compileOnly(project(":database"))
 	compileOnly(project(":shared"))
 	compileOnly("com.github.MilkBowl:VaultAPI:1.7.1")
 }
