+++ plugins/job/src/main/kotlin/fr/pickaria/job/JobPayment.kt
@@ -1,6 +1,6 @@
 package fr.pickaria.job
 
-import fr.pickaria.shared.models.Job
+import fr.pickaria.database.models.Job
 import net.kyori.adventure.text.Component
 import net.milkbowl.vault.economy.EconomyResponse
 import org.bukkit.Sound
