+++ plugins/lib/src/main/kotlin/fr/pickaria/lib/Main.kt
@@ -1,0 +1,5 @@
+package fr.pickaria.lib
+
+import org.bukkit.plugin.java.JavaPlugin
+
+class Main : JavaPlugin()
