+++ plugins/lib/src/main/resources/plugin.yml
@@ -1,0 +1,7 @@
+name: PickariaLib
+version: 1.0
+api-version: 1.19
+main: fr.pickaria.lib.Main
+author: Quozul
+depend:
+  - PickariaShared
