+++ plugins/menu/build.gradle.kts
@@ -1,5 +1,5 @@
 dependencies {
-	compileOnly(project(":shared"))
+	compileOnly(project(":lib"))
 	compileOnly("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4")
 	compileOnly(kotlin("reflect"))
 }
